refactor buildAllQueries
